Rename: services_bundle to redux_bundle
- renamed in test file name
